+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Removed `envvarSplit` parameter from `option()` and `convert()`. Option values from environment variables are no longer split automatically. ([#177](https://github.com/ajalt/clikt/issues/177))
 - Removed public constructors from the following classes: `ProcessedArgument`, `OptionWithValues`, `FlagOption`, `CoOccurringOptionGroup`, `ChoiceGroup`, `MutuallyExclusiveOptions`.
 - `MissingParameter` exception replaced with `MissingOption` and `MissingArgument`
+- Removed `Context.helpOptionMessage`. Override `Localization.helpOptionMessage` and set it on your context instead.
 
 ### Deprecated
 - `@ExperimentalCompletionCandidates` and `@ExperimentalValueSourceApi` annotations. These APIs no longer require an opt-in.